Refactor document generator
